Internationalization

+++ src/navigation/CustomDrawer/CustomDrawer.js
@@ -7,33 +7,36 @@ import {
   Image,
   Dimensions,
   StyleSheet,
-  StatusBar
+  StatusBar,
+  Platform
 } from 'react-native'
 import LinearGradient from 'react-native-linear-gradient'
 import { useDispatch, useSelector } from 'react-redux'
 import { Images } from '../../constants/images'
 import AsyncStorage from '@react-native-async-storage/async-storage'
 import { LOGOUT } from '../../stores/actions/actionType'
+import { useTranslation } from 'react-i18next'
 
 export const CustomDrawer = ({ navigation }) => {
+  const { t } = useTranslation()
   const [active, setActive] = useState('')
   const isGuest = useSelector(state => state.userReducer.isGuest)
   const dispatch = useDispatch()
   const Data = [
     {
-      title: 'Home',
+      title: t('Home'),
       navigateTo: 'Home'
     },
     {
-      title: 'About Us',
+      title: t('About_Us'),
       navigateTo: 'AboutUs'
     },
     {
-      title: 'Settings',
+      title: t('Settings'),
       navigateTo: 'Settings'
     },
     {
-      title: isGuest ? 'Sign Up' : 'Logout',
+      title: isGuest ? t('Sign_Up') : t('LogOut'),
       navigateTo: isGuest ? 'Sign Up' : 'Sign In'
     }
     // {
@@ -43,7 +46,11 @@ export const CustomDrawer = ({ navigation }) => {
   ]
   const removeUser = async () => {
     try {
-      await AsyncStorage.clear()
+      if (Platform.OS == 'ios') {
+        await AsyncStorage.getAllKeys().then(AsyncStorage.multiRemove)
+      } else {
+        await AsyncStorage.clear()
+      }
       dispatch(dispatchh => {
         dispatchh({ type: LOGOUT })
       })
@@ -52,6 +59,7 @@ export const CustomDrawer = ({ navigation }) => {
       })
       return true
     } catch (exception) {
+      console.log('exception exception', exception)
       return false
     }
   }
